Allow all carriers to access BlockedNumberProvider.
The previous permission check only allowed carrier apps for the
default subscription to access the provider. However, the provider is
a singleton, with numbers blocked globally rather than on a
per-subscription basis. Therefore, even non-default carrier apps
should be able to access the database, as such carriers can still be
in active use in DSDS mode.

Fixes: 133232808
Test: TreeHugger to run unit test
Change-Id: I68811f4032a452e0b3478bb5f4554f71fdc66067